backend/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.2.1
psycopg2-binary==2.9.11
